% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{add_to_cache,Readabel-method}
\alias{add_to_cache,Readabel-method}
\title{Add object to cache of x under name}
\usage{
\S4method{add_to_cache}{Readabel}(x, name, object)
}
\arguments{
\item{x}{An object of class Readabel}

\item{name}{A column name}

\item{object}{A column vector of x}
}
\description{
Add object to cache of x under name
}

