% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{as.data.frame,Readabel-method}
\alias{as.data.frame,Readabel-method}
\title{Coerce a Readabel object into a data.frame}
\usage{
\S4method{as.data.frame}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Coerce a Readabel object into a data.frame
}

