% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{$,Readabel-method}
\alias{$,Readabel-method}
\title{Return a given column}
\usage{
\S4method{$}{Readabel}(x, name)
}
\arguments{
\item{x}{An object of class Readabel}

\item{name}{A column name}
}
\description{
Return a given column
}

