% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{clear_cache,Readabel-method}
\alias{clear_cache,Readabel-method}
\title{Clear cache of x}
\usage{
\S4method{clear_cache}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Clear cache of x
}

