% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{colnames,Readabel-method}
\alias{colnames,Readabel-method}
\title{Return the column names of a Readabel object}
\usage{
\S4method{colnames}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Return the column names of a Readabel object
}

