% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_files.R
\name{create_fake_layout_file}
\alias{create_fake_layout_file}
\title{Create a fake OmicABEL layout file}
\usage{
create_fake_layout_file(filename)
}
\arguments{
\item{filename}{The path to the file.}
}
\description{
Create a fake OmicABEL layout file
}

