% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readabel-class.R
\docType{methods}
\name{dim,Readabel-method}
\alias{dim,Readabel-method}
\title{Return dimensions}
\usage{
\S4method{dim}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Return dimensions
}

