% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{dimnames,Readabel-method}
\alias{dimnames,Readabel-method}
\title{Return set of dimnames of a Readabel object}
\usage{
\S4method{dimnames}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Return set of dimnames of a Readabel object
}

