% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{find_in_cache,Readabel-method}
\alias{find_in_cache,Readabel-method}
\title{Find name in cache of x}
\usage{
\S4method{find_in_cache}{Readabel}(x, name)
}
\arguments{
\item{x}{An object of class Readabel}

\item{name}{A column name}
}
\description{
Find name in cache of x
}

