% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{head,Readabel-method}
\alias{head,Readabel-method}
\title{Return the first few rows of a Readabel object}
\usage{
\S4method{head}{Readabel}(x, n = 6L)
}
\arguments{
\item{x}{An object of class Readabel}

\item{n}{Number of rows to return}
}
\description{
Return the first few rows of a Readabel object
}

