% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readabel-class.R
\docType{methods}
\name{initialize,Readabel-method}
\alias{initialize,Readabel-method}
\title{Return a new Readabel object.}
\usage{
\S4method{initialize}{Readabel}(.Object, filename)
}
\arguments{
\item{filename}{The path to a binary OmicABEL output file}

\item{x}{An uninitialized object of class Readabel}
}
\description{
Return a new Readabel object.
}

