% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{is_in_cache,Readabel-method}
\alias{is_in_cache,Readabel-method}
\title{Is a given column in the cache of x?}
\usage{
\S4method{is_in_cache}{Readabel}(x, name)
}
\arguments{
\item{x}{An object of class Readabel}

\item{name}{A column name}
}
\description{
Is a given column in the cache of x?
}

