% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{names,Readabel-method}
\alias{names,Readabel-method}
\title{Return column names.}
\usage{
\S4method{names}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel.}
}
\description{
Return column names.
}

