% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readabel-class.R
\docType{methods}
\name{ncol,Readabel-method}
\alias{ncol,Readabel-method}
\title{Return number of columns.}
\usage{
\S4method{ncol}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Return number of columns.
}

