% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{nrow,Readabel-method}
\alias{nrow,Readabel-method}
\title{Return number of rows.}
\usage{
\S4method{nrow}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Return number of rows.
}

