% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{print,Readabel-method}
\alias{print,Readabel-method}
\title{Print representation of a Readabel object}
\usage{
\S4method{print}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Print representation of a Readabel object
}

