% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\name{read_omicabel}
\alias{read_omicabel}
\title{Read a binary OmicABEL file.}
\usage{
read_omicabel(layout_file, data_file)
}
\arguments{
\item{layout_file}{Path to layout file.}

\item{data_file}{Path to data file.}
}
\description{
Read a binary OmicABEL file.
}

