% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{class}
\name{Readabel-class}
\alias{Readabel-class}
\title{Proxy of a binary OmicABEL output file.}
\description{
Proxy of a binary OmicABEL output file.
}
\section{Slots}{

\describe{
\item{\code{pointer}}{External pointer to a C++ object.}
}}

