% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{row.names,Readabel-method}
\alias{row.names,Readabel-method}
\title{Return the row names of a Readabel object}
\usage{
\S4method{row.names}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Return the row names of a Readabel object
}

