% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{show,Readabel-method}
\alias{show,Readabel-method}
\title{Print representation of a Readabel object}
\usage{
\S4method{show}{Readabel}(object)
}
\arguments{
\item{object}{An object of class Readabel}
}
\description{
Print representation of a Readabel object
}

