% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{snpNames,Readabel-method}
\alias{snpNames,Readabel-method}
\title{Return snp labels}
\usage{
\S4method{snpNames}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Return snp labels
}

