% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readabel-class.R
\docType{methods}
\name{traitNames,Readabel-method}
\alias{traitNames,Readabel-method}
\title{Return trait labels}
\usage{
\S4method{traitNames}{Readabel}(x)
}
\arguments{
\item{x}{An object of class Readabel}
}
\description{
Return trait labels
}

