% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\docType{methods}
\name{initialize,Readabel-method}
\alias{initialize,Readabel-method}
\title{Return a new Readabel object.}
\usage{
\S4method{initialize}{Readabel}(.Object, layout_file, data_file)
}
\arguments{
\item{layout_file}{Path to layout file}

\item{data_file}{Path to data file}

\item{x}{An uninitialized object of class Readabel}
}
\description{
Return a new Readabel object.
}

