% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readabel-class.R
\name{read_omicabel}
\alias{read_omicabel}
\title{Read a binary OmicABEL file.}
\usage{
read_omicabel(filename)
}
\arguments{
\item{filename}{The path to the file.}
}
\description{
Read a binary OmicABEL file.
}

